Создание отдельных классов Model и AjaxController для аддонов
