+++ Field/Addon/Model.php
@@ -1,0 +1,162 @@
+<?php
+/**
+ * Ideal CMS (http://idealcms.ru/)
+ *
+ * @link      http://github.com/ideals/idealcms репозиторий исходного кода
+ * @copyright Copyright (c) 2012-2014 Ideal CMS (http://idealcms.ru/)
+ * @license   http://idealcms.ru/license.html LGPL v3
+ */
+
+namespace Ideal\Field\Addon;
+
+use Ideal\Core\Config;
+use Ideal\Core\Util;
+
+class Model
+{
+    /** @var  string Название поля, используемое для полей ввода в html-коде */
+    public $htmlName;
+
+    /** @var  array Параметры поля, взятые из конфигурационного файла структуры */
+    protected $field;
+
+    /** @var  string Название вкладки, в которой находится поле в окне редактирования */
+    protected $groupName;
+
+    /** @var  \Ideal\Core\Admin\Model Модель данных, в которой находится редактируемое поле */
+    protected $model;
+
+    /** @var  string Название поля */
+    protected $name;
+
+    /**
+     * Получение списка подключённых аддонов для их редактирования
+     *
+     * @return string
+     */
+    public function getAvailableAddonsList()
+    {
+        // Скрываем выбор добавляемого аддона за кнопкой +
+        $html = '<button id="add-addon-button" class="btn btn-default">+</button>'
+            . '<div id="add-addon" class="input-group hide">'
+            . '<select class="form-control" name="add-addon-select" id="add-addon-select">';
+
+        // Получаем список доступных для добавления аддонов для этого элемента
+        $className = $this->field['medium'];
+        /** @var \Ideal\Medium\AbstractModel $medium */
+        $medium = new $className($this->model, $this->name);
+        $list = $medium->getList();
+
+        foreach ($list as $k => $v) {
+            $html .= '<option value="' . $k . '">' . $v . '</option>';
+        }
+        $html .= '</select>';
+
+        // Кнопка добавления аддона, после его выбора в select
+        $html .= '<span class="input-group-btn">'
+            . '<button type="button" id="add-addon-add" class="btn btn-default">+</button>'
+            . '<button type="button" id="add-addon-hide" class="btn btn-default">&times;</button>'
+            . '</span>'
+            . '</div>';
+
+        return $html;
+    }
+
+    /**
+     * Получения содержимого всех вкладок для первоначального отображения
+     *
+     * @param $json
+     * @return array
+     */
+    public function getTabs($json)
+    {
+        $arr = json_decode($json);
+        $result = array(
+            'names' => '',
+            'contents' => ''
+        );
+
+        foreach ($arr as $v) {
+            $tabId = $v[0];
+            $addonVar = $v[1];
+            $addonName = $v[2];
+
+            // Получаем название и содержимое вкладки
+            $tab = $this->getTab($tabId, $addonVar);
+
+            $addonVar .= '_' . $tabId;
+            $addonName = ($addonName == '') ? $tab['name'] : $addonName; // если по какой-то причине в БД сбросится
+
+            // Записываем заголовок вкладки
+            $result['names'] .= addslashes(
+                "<!--suppress HtmlUnknownAnchorTarget -->" // это чтобы PhpStorm не ругался на href='#tab{$addonVar}'
+                . "<li><a data-toggle=\"tab\" id=\"tab{$addonVar}Head\" href='#tab{$addonVar}'>{$addonName}</a>"
+                . "</li>"
+            );
+
+            // Записываем содержимое вкладки
+            $result['contents'] .= "<div id=\"tab{$addonVar}\" class=\"tab-pane\">{$tab['content']}<\/div>";
+        }
+
+        return $result;
+    }
+
+    /**
+     * Получение названия и содержимого одной вкладки
+     *
+     * @param  integer $id Идентификатор вкладки
+     * @param string $addonName Название аддона
+     * @return array
+     */
+    protected function getTab($id, $addonName)
+    {
+        $class = Util::getClassName($addonName, 'Addon') . '\\Model';
+        /** @var \Ideal\Core\Admin\Model $model */
+        $model = new $class('');
+        $model->setFieldsGroup($this->name . '_' . $id);
+        // Загрузка данных связанного объекта
+        $pageData = $this->model->getPageData();
+        if (isset($pageData['ID'])) {
+            $config = Config::getInstance();
+            $path = $this->model->getPath();
+            $end = end($path);
+            $prevStructure = $config->getStructureByName($end['structure']);
+            $prevStructure = $prevStructure['ID'] . '-' . $pageData['ID'];
+            $model->setPageDataByPrevStructure($prevStructure);
+        }
+
+        // Получение содержимого вкладки
+        $tabContent = $model->getFieldsList($model->fields);
+
+        $tabContent = str_replace(
+            array("\\", '<script>', '</script>', "'"),
+            array("\\\\", '\<script>', '<\/script>', "\\'"),
+            $tabContent
+        );
+
+        // Убираем переводы строки, иначе текст не обрабатывается в JS
+        $tabContent = str_replace(array("\n\r", "\r\n", "\n", "\r"), '\\n', $tabContent);
+
+        return array('name' => $model->params['name'], 'content' => $tabContent);
+    }
+
+    /**
+     * Установка модели редактируемого объекта, частью которого является редактируемое поле
+     *
+     * Полю необходимо получать сведения о состоянии объекта и о других полях, т.к.
+     * его значения и поведение может зависеть от значений других полей
+     *
+     * @param \Ideal\Core\Admin\Model $model     Модель редактируемого объекта
+     * @param string                  $fieldName Редактируемое поле
+     * @param string                  $groupName Вкладка, к которой принадлежит редактируемое поле
+     */
+    public function setModel($model, $fieldName, $groupName = 'general')
+    {
+        $this->name = $fieldName;
+        $this->model = $model;
+        $this->field = $model->fields[$fieldName];
+        $this->groupName = $groupName;
+        $this->htmlName = $this->groupName . '_' . $this->name;
+    }
+
+}
